Add PHP 8.4 compatibility
Updated dependencies, tests, and configurations to ensure support for PHP 8.4. Refactored test classes and improved error handling in cURL and stream operations for enhanced robustness. Added Gitpod configuration and VSCode launch settings for better development environment support.
