+++ .vscode/launch.json
@@ -1,0 +1,35 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+		{  
+			"name": "Debug current Script in Console",  
+			"type": "php",  
+			"request": "launch",  
+			"program": "${file}",  
+			"cwd": "${fileDirname}",  
+			"port": 9003,  
+			"runtimeArgs": [  
+				"-dxdebug.start_with_request=yes"  
+			],  
+			"env": {  
+				"XDEBUG_MODE": "debug,develop",  
+				"XDEBUG_CONFIG": "client_port=${port}"  
+			}  
+		},
+		{
+            "name": "PHPUnit Debug",
+            "type": "php",
+            "request": "launch",
+            "program": "${workspaceFolder}/vendor/bin/phpunit",
+            "cwd": "${workspaceFolder}",
+            "port": 9003,
+            "runtimeArgs": [
+                "-dxdebug.start_with_request=yes"
+            ],
+            "env": {
+                "XDEBUG_MODE": "debug,develop",
+                "XDEBUG_CONFIG": "client_port=${port}"
+            }
+        }
+    ]
+}
