+++ composer.json
@@ -20,7 +20,7 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": ">=8.1 <8.4",
+    "php": ">=8.1 <8.5",
     "ext-curl": "*",
     "ext-json": "*",
     "byjg/uri": "^5.0",
@@ -29,8 +29,8 @@
     "psr/http-factory": "^1.0|^1.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.6",
-    "vimeo/psalm": "^5.9",
+    "phpunit/phpunit": "^10.5|^11.5",
+    "vimeo/psalm": "^5.9|^6.2",
     "ext-soap": "*"
   },
   "provide": {
