+++ tests/Psr7/FileStreamAPlusTestStreamBase.php
@@ -2,15 +2,20 @@
 
 namespace Test\Psr7;
 
-use ByJG\WebRequest\Factory\StreamFactory;
-use ByJG\WebRequest\Psr7\TempFileStream;
+use ByJG\WebRequest\Psr7\FileStream;
 use Psr\Http\Message\StreamInterface;
 
-class TempFileStreamTest extends StreamBaseTest
+class FileStreamAPlusTestStreamBase extends TestStreamBase
 {
+    const FILENAME = "/tmp/filestream-test.txt";
+
     public function getResource(?string $data): StreamInterface
     {
-        return new TempFileStream($data);
+        if (file_exists(self::FILENAME)) {
+            unlink(self::FILENAME);
+        }
+        file_put_contents(self::FILENAME, $data);
+        return new FileStream(self::FILENAME, "a+");
     }
 
     /**
@@ -20,8 +25,8 @@ class TempFileStreamTest extends StreamBaseTest
     {
         $this->stream->close();
         $this->stream = null;
+        unlink(self::FILENAME);
     }
-
     /**
      * @return true
      */
@@ -31,17 +36,10 @@ class TempFileStreamTest extends StreamBaseTest
     }
 
     /**
-     * @return true
+     * @return false
      */
     public function canOverwrite()
     {
-        return true;
-    }
-
-    public function testCreateStream()
-    {
-        $stream = StreamFactory::instance(TempFileStream::class)->createStream("test");
-
-        $this->assertEquals("test", (string)$stream);
+        return false;
     }
 }
