+++ tests/Psr7/TestStreamBase.php
@@ -5,7 +5,7 @@ namespace Test\Psr7;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\StreamInterface;
 
-abstract class StreamBaseTest extends TestCase
+abstract class TestStreamBase extends TestCase
 {
     const TEXT1 = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam auctor augue justo, id condimentum tortor molestie et. Quisque at egestas dui. Vestibulum id lectus et mi interdum lobortis non sit.";
